+++ ReleaseNotes.md
@@ -1,6 +1,12 @@
 
 # Release Notes
 
+## Version 1.5.1
+
+### User-visible Changes
+
+* The Bomber Jacket item no shows a bomb range highlight.
+
 ## Version 1.5.0
 
 ### User-visible Changes